Delete ReproducibleR_AssignmentWork.Rmd
I have edited a few spelling and grammar points on my main file since uploading it to Github, which I would like reflected in the file here, so I am deleting this file and reuploading it again. None of the changes are major differences to how the file was when it was reviewed by my partners and I have not changed any of the Q2 code from how it was at the time of review, so the reviews I received are still exactly relevant to the newly uploaded markdown file.